requirements.txt
Babel==2.7.0
Django==2.2.2
django-uuslug==1.1.8
Pillow==6.0.0
pkg-resources==0.0.0
python-slugify==3.0.2
pytz==2019.1
sqlparse==0.3.0
text-unidecode==1.2
